requirements_dev.txt
numpy
cartopy
black
dateparser
docker
flake8
geos
gitpython
h5py
matplotlib
mercantile
mypy
obspy
packaging
pandas
pycsep
pydocstringformatter
pyproj
pyshp
pytest
pytest-cov
pyyaml
requests
scipy
seaborn
shapely
sphinx
sphinx-autoapi
sphinx_design
sphinx-gallery
sphinx-rtd-theme
sphinx-toolbox
sphinx_copybutton
tables
tox
vcrpy==4.3.1
xmltodict